Moving over from having the map responsible for Camera to a camera obj
I feel like this is making it easier, at least I don't need to worry
about adding even more stuff to my map, so that's a plus.

There is still an issue with how things are rendered on the lower
right-hand side because I can't math apparently. Basically when I get
to that edge of the map my calculations just shit the bed. It might
be worth just moving back to using X and Y values. It's more verbose
but I won't be a victim to my crappy point logic that I built.